services/face-collector/requirements.txt
paho-mqtt
sqlite3